Layout & Scaling Improvements and bug fixes
I refactored all of the code to use the Pos module to handle scaling and resizing. This layout can now be run on 4:3, 16:9 or even vertically. I also added and changed some settings, and added a few more visual details. With the new code, the videos now display correctly when shown on larger screens.